Flask App/requirements.txt
# Automatically generated by https://github.com/damnever/pigar.

# F:\ai2004_finale_proj\flask_app\app.py: 3
Flask == 1.1.2

# F:\ai2004_finale_proj\flask_app\app.py: 5,7,8
Keras == 2.4.3

# F:\ai2004_finale_proj\flask_app\app.py: 6
Pillow == 7.2.0

# F:\ai2004_finale_proj\flask_app\app.py: 4
Werkzeug == 1.0.1

# F:\ai2004_finale_proj\flask_app\app.py: 9
numpy == 1.17.0
